Moved fp_sensitivity etc. to basis/_helpers.py
The `_functions.py` module is meant for things that
will be called by the wrapper function
